add components
Chat and update of App
